+++ src/index.js
@@ -1,6 +1,6 @@
 import '@babel/polyfill';
 import Vue from 'vue';
-import App from './App.vue';
+import App from './components/App.vue';
 import './assets/favicon.ico';
 import './assets/manifest.json';
